dog_backend/requirements.txt
Flask
flask-cors
psycopg2-binary
tensorflow
Pillow
matplotlib
numpy
pandas
lime
waitress
